dry up teacher and assistant teacher classes with teach module
